requirements.txt
google-generativeai
python-dotenv
rapidfuzz
ranker